api/app/requirements.txt
uvicorn~=0.25.0
python-ffmpeg~=0.2.0
fastapi~=0.108.0
SQLAlchemy~=2.0.25
pydantic~=2.5.3
cryptography~=41.0.7
pydantic-settings~=2.1.0
PyJWT~=2.8.0
